commands: prettier category and rule printing in `mllint describe` and `mllint list`, moves prettyPrint functions to `commands/util.go`
